+++ CHANGELOG.md
@@ -53,6 +53,7 @@ The following changes have been completed, and will be included in the next tagg
   - Added `URLFormatOptionsWithHash` dumb value type
 * URL
   - Added `buildURLHref()`
+  - Added `isURLSearch()` data guard
   - Added `URL` value type
 
 ### Fixes